services/python/pyproject.toml
[project]
name = "rapidmq-python"
version = "0.1.0"
description = "RapidMQ Python example"
requires-python = ">=3.11"

dependencies = [
    "aio-pika>=9.4",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
]

[dependency-groups]
dev = ["ruff>=0.15.11"]

[project.scripts]
rapidmq-python = "main:run"
